# src/swagger/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Gym App API",
    "version": "1.0.0",
    "description": "API for Gym App"
  },
  "host": "localhost:3000",
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/exercise/{uuid}": {
      "get": {
        "tags": [
          "Exercise"
        ],
        "summary": "Get an exercise by ID",
        "description": "Retrieves an exercise by its ID.",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Exercise ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Exercise found."
          },
          "404": {
            "description": "Exercise not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/exercise/name/{name}": {
      "get": {
        "tags": [
          "Exercise"
        ],
        "summary": "Get an exercise by name",
        "description": "Retrieves an exercise by its name.",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Exercise name"
          }
        ],
        "responses": {
          "200": {
            "description": "Exercise found."
          },
          "404": {
            "description": "Exercise not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
...
    "/users/create": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Create a new user",
        "description": "Creates a new user.",
        "parameters": [
          {
            "name": "username",
            "description": "User username",
            "in": "query",
            "type": "string"
          },
          {
            "name": "email",
            "description": "User email",
            "in": "query",
            "type": "string"
          },
          {
            "name": "password",
            "description": "User password",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "User created."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/workout/delete/{uuid}": {
      "delete": {
        "tags": [
          "Workout"
        ],
        "summary": "Remove workout by workout ID",
        "description": "Removes a workout by its ID.",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Workout ID"
          }
        ],
        "responses": {
          "204": {
            "description": "Workout removed."
          },
          "404": {
            "description": "Workout not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/workout/delete/{uuid}/exercise": {
      "delete": {
        "tags": [
          "Workout"
        ],
        "summary": "Remove exercises from workout by workout ID and exercise ID",
        "description": "Removes exercises from a workout by its ID.",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Workout ID"
          },
          {
            "name": "exercises",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "Exercises removed."
          },
          "404": {
            "description": "Workout not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/workout/delete/{uuid}/exercise/all": {
      "delete": {
        "tags": [
          "Workout"
        ],
        "summary": "Remove all exercises from workout by workout ID",
        "description": "Removes all exercises from a workout by its ID.",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Workout ID"
          }
        ],
        "responses": {
          "204": {
            "description": "Exercises removed."
          },
          "404": {
            "description": "Workout not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/workout/all": {
      "get": {
        "tags": [
          "Workout"
        ],
        "summary": "Get all workouts",
        "description": "Retrieves all workouts.",
        "responses": {
          "200": {
            "description": "Workouts found."
          },
          "404": {
            "description": "Workouts not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/workout/all/exercises": {
      "get": {
        "tags": [
          "Workout"
        ],
        "summary": "Get workouts and exercises",
        "description": "Retrieves workouts and exercises.",
        "responses": {
          "200": {
            "description": "Workouts found."
          },
          "404": {
            "description": "Workouts not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/workout/{uuid}": {
      "get": {
        "tags": [
          "Workout"
        ],
        "summary": "Get workout by workout ID",
        "description": "Retrieves a workout by its ID.",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Workout ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Workout found."
          },
          "404": {
            "description": "Workout not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/workout/{uuid}/exercises": {
      "get": {
        "tags": [
          "Workout"
        ],
        "summary": "Get workout exercises by workout ID",
        "description": "Retrieves a workout's exercises by its ID.",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Workout ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Workout found."
          },
          "404": {
            "description": "Workout not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/workout/name/{name}": {
      "get": {
        "tags": [
          "Workout"
        ],
        "summary": "Get workout by name",
        "description": "Retrieves a workout by its name.",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Workout name"
          }
        ],
        "responses": {
          "200": {
            "description": "Workout found."
          },
          "404": {
            "description": "Workout not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/workout/exercise/{uuid}": {
      "get": {
        "tags": [
          "Workout"
        ],
        "summary": "Get workouts by exercise ID",
        "description": "Retrieves workouts by exercise ID.",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Exercise ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Workouts found."
          },
          "404": {
            "description": "Workouts not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/workout/exercise/name/{name}": {
      "get": {
        "tags": [
          "Workout"
        ],
        "summary": "Get workouts by exercise name",
        "description": "Retrieves workouts by exercise name.",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Exercise name"
          }
        ],
        "responses": {
          "200": {
            "description": "Workouts found."
          },
          "404": {
            "description": "Workouts not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/workout/update/{uuid}": {
      "put": {
        "tags": [
          "Workout"
        ],
        "summary": "Update workout by workout ID",
        "description": "Updates a workout by its ID.",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Workout ID"
          },
          {
            "name": "name",
            "description": "Workout name",
            "in": "query",
            "type": "string"
          },
          {
            "name": "exercises",
            "description": "Workout exercises",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Workout updated."
          },
          "404": {
            "description": "Workout not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    }
  }
}